+++ physic_engine/force.c
@@ -15,7 +15,7 @@ vector statsys_bounce_force(
     statsys const *s,
     dynpoint const *m,
     vector const *contact_normal,
-    float dt
+    double dt
 ) {
     return vector__mul(
         contact_normal,
@@ -28,7 +28,7 @@ vector dynpoint_bounce_force(
     dynpoint const *m1,
     dynpoint const *m2,
     vector const *contact_normal,
-    float dt
+    double dt
 ) {
     UNUSED(m1);
     UNUSED(m2);
