+++ physic_engine/physic_object.h
@@ -13,19 +13,19 @@ typedef struct
     vector acceleration;
     vector speed;
     vector position;
-    float mass;
+    double mass;
     sphere collision_sphere;
 } dynpoint;
 
 dynpoint *dynpoint__new(
     vector initial_position,
     vector initial_speed,
-    float radius,
-    float mass
+    double radius,
+    double mass
 );
 void dynpoint__delete(dynpoint *system);
 
-void dynpoint__apply_forces(dynpoint *m, vector const *f, float dt);
+void dynpoint__apply_forces(dynpoint *m, vector const *f, double dt);
 void dynpoint__print(dynpoint *system);
 
 /**
@@ -33,13 +33,13 @@ void dynpoint__print(dynpoint *system);
  */
 typedef struct
 {
-    float bounce_coeff;
+    double bounce_coeff;
     vector position;
     geometric_shape
         *shape; // shape used to perform collision detection on the system
 } statsys;
 
-statsys *bouncer__new(vector position, float radius);
+statsys *bouncer__new(vector position, double radius);
 statsys *wall__new(vector position, vector normal);
 void statsys__delete(statsys *system);
